Added support for smart bind mounting of general libs and some general fixes
On 7218 platform some general libs are bind mounted into the container because
the graphic libs depend on them. But the you.i application depends on newer
versions of those same libraries so the application was not working with the
old version from the 7218 host platform. We now check if the general libs
delivered by the container are OK for the graphic libs from the platform.

* split of libs.txt in libs.txt (graphics libs) and genlibs.txt (general libs)
* added some extra .x version of the libs in 7218 and 7218refapp
* added libs_deps.txt in 7218 platform with information on the needed version of the general libs
* added scripting support the check and include only the general libs from host platform when needed by dependencies
* fixed the scp_client.sh and pull_client.sh so they can be used on boxes with changing host keys